requirements.txt
pyyaml>=6.0
requests>=2.31.0
flask>=3.0
python-crontab>=3.0

